+++ Sources/WalletConnectJWT/Claims/RelayClaims.swift
@@ -4,6 +4,6 @@ public struct RelayClaims: JWTEncodable {
     let iss: String
     let sub: String
     let aud: String
-    let iat: Int
-    let exp: Int
+    let iat: Int64
+    let exp: Int64
 }
